Clean up php3.*\.h files. The files itself are renamed, and references in all .*\.[ch] files were changed. There is a slight chance that my script missed a few changes, please correct them manually.

+++ ext/pgsql/pgsql.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
  */
  
-/* $Id: pgsql.c,v 1.27 1999/11/20 01:14:36 andi Exp $ */
+/* $Id: pgsql.c,v 1.28 1999/12/04 19:16:00 sas Exp $ */
 
 #include <stdlib.h>
 
@@ -26,8 +26,8 @@
 #endif
 
 #include "php.h"
-#include "php3_pgsql.h"
-#include "ext/standard/php3_standard.h"
+#include "php_pgsql.h"
+#include "ext/standard/php_standard.h"
 #include "php_globals.h"
 
 #if HAVE_PGSQL
@@ -331,7 +331,7 @@ void php3_pgsql_do_connect(INTERNAL_FUNCTION_PARAMETERS,int persistent)
 			if (index_ptr->type != le_index_ptr) {
 				RETURN_FALSE;
 			}
-			link = (int) index_ptr->ptr;
+			link = (int) (long) index_ptr->ptr;
 			ptr = php3_list_find(link,&type);   /* check if the link is still there */
 			if (ptr && (type==le_link || type==le_plink)) {
 				return_value->value.lval = PGG(default_link) = link;
